Creating API endpoint for upvoting and downvoting evaluations
The endpoint is /api/evaluations/EVALUATION_ID/votes. Use the HTTP verb
PUT to up vote, and DELETE to downvote. For example, to upvote in
Javascript:
$.ajax('/api/evaluations/58e139a8467b099df7953e75/votes', {method:
'PUT’})

If a user tries to do an action twice (such as upvoting the same
evaluation multiple times an error 403 (Forbidden) will be returned.