Improving score shown in search results
For courses for which evaluations are not (yet) available, the score
shown in search results will now show the most recent semester taught
by the instructor teaching the course in the semester being searched
for. If this instructor has never taught the course, then the score for
the most recent semester will be shown.